fix: use bg location to open in new tab. share modal bg logic
