+++ src/app/pages/home/home.tsx
@@ -1,11 +1,9 @@
 import { Route, useNavigate } from 'react-router-dom';
-import { Outlet, Routes, useLocation } from 'react-router-dom';
 
 import { RouteUrls } from '@shared/route-urls';
 
 import { useTrackFirstDeposit } from '@app/common/hooks/analytics/transactions-analytics.hooks';
 import { useOnboardingState } from '@app/common/hooks/auth/use-onboarding-state';
-import { useLocationState } from '@app/common/hooks/use-location-state';
 import { useOnMount } from '@app/common/hooks/use-on-mount';
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
 import { Header } from '@app/components/header';
@@ -13,6 +11,7 @@ import { ActivityList } from '@app/features/activity-list/activity-list';
 import { AssetsList } from '@app/features/asset-list/asset-list';
 import { InAppMessages } from '@app/features/hiro-messages/in-app-messages';
 import { homePageModalRoutes } from '@app/routes/app-routes';
+import { ModalBackgroundWrapper } from '@app/routes/components/modal-background-wrapper';
 
 import { CurrentAccount } from './components/account-area';
 import { HomeTabs } from './components/home-tabs';
@@ -21,8 +20,6 @@ import { HomeLayout } from './components/home.layout';
 export function Home() {
   const { decodedAuthRequest } = useOnboardingState();
 
-  const location = useLocation();
-  const backgroundLocation = useLocationState<Location>('backgroundLocation');
   const navigate = useNavigate();
 
   useTrackFirstDeposit();
@@ -41,22 +38,14 @@ export function Home() {
   return (
     <HomeLayout currentAccount={<CurrentAccount />}>
       <HomeTabs>
-        <>
-          {/* 
-          To overlay modal on nested routes backgroundLocation is used 
-          to trick the router into thinking its on the same page
-          */}
-          <Routes location={backgroundLocation || location}>
-            <Route index element={<AssetsList />} />
-            <Route path={RouteUrls.Activity} element={<ActivityList />}>
-              {homePageModalRoutes}
-            </Route>
-
+        <ModalBackgroundWrapper>
+          <Route index element={<AssetsList />} />
+          <Route path={RouteUrls.Activity} element={<ActivityList />}>
             {homePageModalRoutes}
-            <Route path="*" element={<Outlet />} />
-          </Routes>
-          {backgroundLocation && <Outlet />}
-        </>
+          </Route>
+
+          {homePageModalRoutes}
+        </ModalBackgroundWrapper>
       </HomeTabs>
     </HomeLayout>
   );
