+++ src/app/pages/sign-out-confirm/sign-out-confirm.tsx
@@ -2,9 +2,9 @@ import { useNavigate } from 'react-router-dom';
 
 import { RouteUrls } from '@shared/route-urls';
 
-import { useBackgroundLocationRedirect } from '@app/common/hooks/use-background-location-redirect';
 import { useKeyActions } from '@app/common/hooks/use-key-actions';
 import { useLocationState } from '@app/common/hooks/use-location-state';
+import { useBackgroundLocationRedirect } from '@app/routes/hooks/use-background-location-redirect';
 
 import { SignOutConfirmLayout } from './sign-out-confirm.layout';
