+++ src/app/routes/hooks/use-background-location-redirect.ts
@@ -7,7 +7,7 @@ import { useLocationState } from '@app/common/hooks/use-location-state';
 
 // If routes are accessed directly / opened in new tabs `backgroundLocation` is lost
 // this hook re-directs the users to home then overlays the modal so the BG is the home route
-export function useBackgroundLocationRedirect() {
+export function useBackgroundLocationRedirect(baseUrl = RouteUrls.Home) {
   const { pathname, state } = useLocation();
   const navigate = useNavigate();
   const backgroundLocation = useLocationState('backgroundLocation');
@@ -16,10 +16,10 @@ export function useBackgroundLocationRedirect() {
     void (async () => {
       if (backgroundLocation === undefined) {
         return navigate(pathname, {
-          state: { backgroundLocation: { pathname: RouteUrls.Home }, ...state },
+          state: { backgroundLocation: { pathname: baseUrl }, ...state },
         });
       }
       return false;
     })();
-  }, [backgroundLocation, navigate, pathname, state]);
+  }, [backgroundLocation, baseUrl, navigate, pathname, state]);
 }
